:pencil2: fix: Traveler Entity 수정

+++ src/main/java/nadeuli/dto/request/ExpenseItemRequestDTO.java
@@ -28,7 +28,7 @@ public class ExpenseItemRequestDTO {
     private String payer;
 
     @NotNull
-    private Integer expense;
+    private Long expense;
 
 
 }
